add eventbus backend
