+++ main/lib/idds/agents/common/eventbus/baseeventbusbackend.py
@@ -6,7 +6,7 @@
 # http://www.apache.org/licenses/LICENSE-2.0OA
 #
 # Authors:
-# - Wen Guan, <wen.guan@cern.ch>, 2022
+# - Wen Guan, <wen.guan@cern.ch>, 2023
 
 import logging
 import time
@@ -17,13 +17,13 @@ import uuid
 from .event import StateClaimEvent, EventBusState
 
 
-class LocalEventBusBackend(threading.Thread):
+class BaseEventBusBackend(threading.Thread):
     """
-    Local Event Bus Backend
+    Base Event Bus Backend
     """
 
     def __init__(self, logger=None, **kwargs):
-        super(LocalEventBusBackend, self).__init__()
+        super(BaseEventBusBackend, self).__init__()
         self._id = str(uuid.uuid4())[:8]
         self._state_claim_wait = 60
         self._state_claim = StateClaimEvent(self._id, EventBusState.New, time.time())
@@ -36,6 +36,8 @@ class LocalEventBusBackend(threading.Thread):
         self._lock = threading.RLock()
 
         self.setup_logger(logger)
+
+        self.coordinator = None
 
     def setup_logger(self, logger=None):
         """
@@ -53,27 +55,52 @@ class LocalEventBusBackend(threading.Thread):
         self.graceful_stop.set()
 
     def send(self, event):
-        with self._lock:
-            if event._event_type not in self._events:
-                self._events[event._event_type] = {}
-                self._events_index[event._event_type] = []
-            self._events[event._event_type][event._id] = event
-            self._events_index[event._event_type].append(event._id)
+        if self.get_coordinator():
+            return self.get_coordinator().send(event)
+        else:
+            with self._lock:
+                if event._event_type not in self._events:
+                    self._events[event._event_type] = {}
+                    self._events_index[event._event_type] = []
+                self._events[event._event_type][event._id] = event
+                self._events_index[event._event_type].append(event._id)
 
     def get(self, event_type, wait=0):
-        with self._lock:
-            if event_type in self._events_index and self._events_index[event_type]:
-                event_id = self._events_index[event_type].pop(0)
-                event = self._events[event_type][event_id]
-                del self._events[event_type][event_id]
-                return event
-            return None
+        if self.get_coordinator():
+            return self.get_coordinator().get(event_type, wait)
+        else:
+            with self._lock:
+                if event_type in self._events_index and self._events_index[event_type]:
+                    event_id = self._events_index[event_type].pop(0)
+                    event = self._events[event_type][event_id]
+                    del self._events[event_type][event_id]
+                    return event
+                return None
+
+    def send_report(self, event, status, start_time, end_time, source, result):
+        if self.get_coordinator():
+            return self.get_coordinator().send_report(event, status, start_time, end_time, source, result)
 
     def clean_event(self, event):
         pass
 
     def fail_event(self, event):
         pass
+
+    def set_manager(self, manager):
+        pass
+
+    def get_manager(self):
+        return None
+
+    def set_coordinator(self, coordinator):
+        self.coordinator = coordinator
+
+    def get_coordinator(self):
+        return self.coordinator
+
+    def is_ok(self):
+        return True
 
     def execute(self):
         while not self.graceful_stop.is_set():
